requirements.txt
flask==2.1.2
paho-mqtt==1.6.1